requirements.txt
streamlit
pandas
groq
gspread
google-auth-oauthlib
google-auth
python-dateutil


